+++ src/main/kotlin/io/github/ydwk/ydwk/rest/RestApiClient.kt
@@ -1,0 +1,4 @@
+package io.github.ydwk.ydwk.rest
+
+class RestApiClient {
+}
