+++ src/main/kotlin/io/github/ydwk/ydwk/rest/endpoint/EndPoint.kt
@@ -16,7 +16,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */ 
-package io.github.ydwk.ydwk.rest
+package io.github.ydwk.ydwk.rest.endpoint
 
 import io.github.ydwk.ydwk.YDWKInfo
