+++ src/main/kotlin/io/github/ydwk/ydwk/rest/manager/RestApiManager.kt
@@ -16,8 +16,9 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */ 
-package io.github.ydwk.ydwk.rest
+package io.github.ydwk.ydwk.rest.manager
 
+import io.github.ydwk.ydwk.rest.endpoint.EndPoint
 import io.github.ydwk.ydwk.rest.type.*
 import okhttp3.RequestBody
